Remove plookup_composer module
Double and remove lookup_lineariser

Double and remove plookup_proof.rs

Double and remove lookup_quotient.rs

Double and remove plookup_preprocess.rs

Remove plookup_prover.rs

Double and remove plookup_verifier.rs

Remove and rename PlookupProver

Remove and rename PlookupVerifier

Remove and rename PlookupProof

Remove and rename in quotient_poly.rs

Remove and rename in linearisation_poly.rs

Remove and rename in preprocess.rs

Remove and rename in preprocess.rs

Remove and rename in verifier.rs

Remove and rename in widget/mod.rs

Rename in lookup/mod.rs

Remove and rename in proof_system/mod.rs

Remove and rename in lookup/verifierkey.rs

Remove in curve_addition/mod.rs

Remove and rename in curve_addition/verifierkey.rs

Remove in scalar_mul/fixed_base/mod.rs

Remove and rename in scalar_mul/fixed_base/verifierkey.rs

Remove and rename in logic/verifierkey.rs

Fixup in logic/verifierkey.rs

Fixup 2 in logic/verifierkey.rs

Remove in logic/mod.rs

Remove and rename in widget/permutation

Remove and rename in widget/range

Remove in constraint_system/arithmetic.rs

Rename and remove in widget/arithmetic

Rename in lookup/proverkey.rs
