+++ src/constraint_system/mod.rs
@@ -10,7 +10,6 @@
 //! build, preprocess circuits.
 pub(crate) mod composer;
 pub(crate) mod cs_errors;
-pub(crate) mod plookup_composer;
 pub(crate) mod variable;
 
 /// Simple Arithmetic gates
@@ -27,5 +26,4 @@ pub mod logic;
 pub mod range;
 
 pub use composer::StandardComposer;
-pub use plookup_composer::PlookupComposer;
 pub use variable::{Variable, WireData};
