+++ src/proof_system/preprocess.rs
@@ -30,6 +30,8 @@ pub(crate) struct SelectorPolynomials {
     q_logic: Polynomial,
     q_fixed_group_add: Polynomial,
     q_variable_group_add: Polynomial,
+    q_lookup: Polynomial,
+
     left_sigma: Polynomial,
     right_sigma: Polynomial,
     out_sigma: Polynomial,
@@ -58,6 +60,7 @@ impl StandardComposer {
         self.q_logic.extend(zeroes_scalar.iter());
         self.q_fixed_group_add.extend(zeroes_scalar.iter());
         self.q_variable_group_add.extend(zeroes_scalar.iter());
+        self.q_lookup.extend(zeroes_scalar.iter());
 
         self.w_l.extend(zeroes_var.iter());
         self.w_r.extend(zeroes_var.iter());
@@ -81,6 +84,7 @@ impl StandardComposer {
             && self.q_logic.len() == k
             && self.q_fixed_group_add.len() == k
             && self.q_variable_group_add.len() == k
+            && self.q_lookup.len() == k
             && self.w_l.len() == k
             && self.w_r.len() == k
             && self.w_o.len() == k
@@ -127,6 +131,8 @@ impl StandardComposer {
             domain_4n.coset_fft(&selectors.q_variable_group_add),
             domain_4n,
         );
+        let q_lookup_eval_4n =
+            Evaluations::from_vec_and_domain(domain_4n.coset_fft(&selectors.q_lookup), domain_4n);
 
         let left_sigma_eval_4n =
             Evaluations::from_vec_and_domain(domain_4n.coset_fft(&selectors.left_sigma), domain_4n);
@@ -190,6 +196,11 @@ impl StandardComposer {
             q_variable_group_add: (selectors.q_variable_group_add, q_variable_group_add_eval_4n),
         };
 
+        // Prover key for lookup operations
+        let lookup_prover_key = widget::lookup::ProverKey {
+            q_lookup: (selectors.q_lookup, q_lookup_eval_4n),
+        };
+
         let prover_key = widget::ProverKey {
             n: domain.size(),
             arithmetic: arithmetic_prover_key,
@@ -198,6 +209,7 @@ impl StandardComposer {
             permutation: permutation_prover_key,
             variable_base: curve_addition_prover_key,
             fixed_base: ecc_prover_key,
+            lookup: lookup_prover_key,
             // Compute 4n evaluations for X^n -1
             v_h_coset_4n: domain_4n.compute_vanishing_poly_over_coset(domain.size() as u64),
         };
@@ -221,7 +233,14 @@ impl StandardComposer {
         &mut self,
         commit_key: &CommitKey,
         transcript: &mut Transcript,
-    ) -> Result<(widget::VerifierKey, SelectorPolynomials, EvaluationDomain), Error> {
+    ) -> Result<
+        (
+            widget::VerifierKey,
+            SelectorPolynomials,
+            EvaluationDomain,
+        ),
+        Error,
+    > {
         let domain = EvaluationDomain::new(self.circuit_size())?;
 
         // Check that the length of the wires is consistent.
@@ -243,6 +262,7 @@ impl StandardComposer {
             Polynomial::from_coefficients_slice(&domain.ifft(&self.q_fixed_group_add));
         let q_variable_group_add_poly =
             Polynomial::from_coefficients_slice(&domain.ifft(&self.q_variable_group_add));
+        let q_lookup_poly = Polynomial::from_coefficients_slice(&domain.ifft(&self.q_lookup));
 
         // 2. Compute the sigma polynomials
         let (left_sigma_poly, right_sigma_poly, out_sigma_poly, fourth_sigma_poly) =
@@ -263,6 +283,7 @@ impl StandardComposer {
         let q_variable_group_add_poly_commit = commit_key
             .commit(&q_variable_group_add_poly)
             .unwrap_or_default();
+        let q_lookup_poly_commit = commit_key.commit(&q_lookup_poly).unwrap_or_default();
 
         let left_sigma_poly_commit = commit_key.commit(&left_sigma_poly)?;
         let right_sigma_poly_commit = commit_key.commit(&right_sigma_poly)?;
@@ -298,6 +319,12 @@ impl StandardComposer {
         let curve_addition_verifier_key = widget::ecc::curve_addition::VerifierKey {
             q_variable_group_add: q_variable_group_add_poly_commit,
         };
+
+        // Verifier Key for lookup operations
+        let lookup_verifier_key = widget::lookup::VerifierKey {
+            q_lookup: q_lookup_poly_commit,
+        };
+
         // Verifier Key for permutation argument
         let permutation_verifier_key = widget::permutation::VerifierKey {
             left_sigma: left_sigma_poly_commit,
@@ -314,6 +341,7 @@ impl StandardComposer {
             fixed_base: ecc_verifier_key,
             variable_base: curve_addition_verifier_key,
             permutation: permutation_verifier_key,
+            lookup: lookup_verifier_key,
         };
 
         let selectors = SelectorPolynomials {
@@ -328,6 +356,7 @@ impl StandardComposer {
             q_logic: q_logic_poly,
             q_fixed_group_add: q_fixed_group_add_poly,
             q_variable_group_add: q_variable_group_add_poly,
+            q_lookup: q_lookup_poly,
             left_sigma: left_sigma_poly,
             right_sigma: right_sigma_poly,
             out_sigma: out_sigma_poly,
@@ -348,9 +377,9 @@ mod test {
     #[test]
     /// Tests that the circuit gets padded to the correct length
     /// XXX: We can do this test without dummy_gadget method
-    fn test_pad() {
+    fn test_plookup_pad() {
         let mut composer: StandardComposer = StandardComposer::new();
-        dummy_gadget(100, &mut composer);
+        dummy_gadget_plookup(100, &mut composer);
 
         // Pad the circuit to next power of two
         let next_pow_2 = composer.n.next_power_of_two() as u64;
@@ -368,6 +397,7 @@ mod test {
         assert!(composer.q_logic.len() == size);
         assert!(composer.q_fixed_group_add.len() == size);
         assert!(composer.q_variable_group_add.len() == size);
+        assert!(composer.q_lookup.len() == size);
         assert!(composer.w_l.len() == size);
         assert!(composer.w_r.len() == size);
         assert!(composer.w_o.len() == size);
