+++ src/proof_system/proof.rs
@@ -11,12 +11,11 @@
 //! `Proof` structure and it's methods.
 
 use super::linearisation_poly::ProofEvaluations;
-use super::lookup_lineariser::PlookupProofEvaluations;
 use super::proof_system_errors::ProofErrors;
 use crate::commitment_scheme::kzg10::{AggregateProof, Commitment, OpeningKey};
 use crate::fft::{EvaluationDomain, Polynomial};
-use crate::plookup::{MultiSet, PreprocessedTable4Arity};
-use crate::proof_system::widget::{PlookupVerifierKey, VerifierKey};
+use crate::plookup::{MultiSet, PlookupTable4Arity, PreprocessedTable4Arity};
+use crate::proof_system::widget::VerifierKey;
 use crate::transcript::TranscriptProtocol;
 use anyhow::{Error, Result};
 use dusk_bls12_381::{multiscalar_mul::msm_variable_base, BlsScalar, G1Affine};
@@ -49,8 +48,20 @@ pub struct Proof {
     /// Commitment to the witness polynomial for the fourth wires.
     pub d_comm: Commitment,
 
+    /// Commitment to the lookup query polynomial.
+    pub f_comm: Commitment,
+
+    /// Commitment to first half of sorted polynomial
+    pub h_1_comm: Commitment,
+
+    /// Commitment to second half of sorted polynomial
+    pub h_2_comm: Commitment,
+
     /// Commitment to the permutation polynomial.
     pub z_comm: Commitment,
+
+    /// Commitment to the plookup permutation polynomial.
+    pub p_comm: Commitment,
 
     /// Commitment to the quotient polynomial.
     pub t_1_comm: Commitment,
@@ -65,6 +76,7 @@ pub struct Proof {
     pub w_z_comm: Commitment,
     /// Commitment to the shifted opening polynomial.
     pub w_zw_comm: Commitment,
+
     /// Subset of all of the evaluations added to the proof.
     pub evaluations: ProofEvaluations,
 }
@@ -161,10 +173,10 @@ impl Proof {
         verifier_key: &VerifierKey,
         transcript: &mut Transcript,
         opening_key: &OpeningKey,
+        lookup_table: &PlookupTable4Arity,
         pub_inputs: &[BlsScalar],
     ) -> Result<(), Error> {
         let domain = EvaluationDomain::new(verifier_key.n)?;
-
         // Subgroup checks are done when the proof is deserialised.
 
         // In order for the Verifier and Prover to have the same view in the non-interactive setting
@@ -178,12 +190,33 @@ impl Proof {
         transcript.append_commitment(b"w_o", &self.c_comm);
         transcript.append_commitment(b"w_4", &self.d_comm);
 
+        // Compute zeta compression challenge
+        let zeta = transcript.challenge_scalar(b"zeta");
+
+        // Add f_poly commitment to transcript
+        transcript.append_commitment(b"f", &self.f_comm);
+
         // Compute beta and gamma challenges
         let beta = transcript.challenge_scalar(b"beta");
         transcript.append_scalar(b"beta", &beta);
         let gamma = transcript.challenge_scalar(b"gamma");
+
+        // Compute delta and epsilon challenges
+        let delta = transcript.challenge_scalar(b"delta");
+        let epsilon = transcript.challenge_scalar(b"epsilon");
+
         // Add commitment to permutation polynomial to transcript
         transcript.append_commitment(b"z", &self.z_comm);
+
+        // Compute evaluation challenge
+        let z_challenge = transcript.challenge_scalar(b"z_challenge");
+
+        // Add h polynomials to transcript
+        transcript.append_commitment(b"h1", &self.h_1_comm);
+        transcript.append_commitment(b"h2", &self.h_2_comm);
+
+        // Add permutation polynomial commitment to transcript
+        transcript.append_commitment(b"p", &self.p_comm);
 
         // Compute quotient challenge
         let alpha = transcript.challenge_scalar(b"alpha");
@@ -193,6 +226,7 @@ impl Proof {
             transcript.challenge_scalar(b"fixed base separation challenge");
         let var_base_sep_challenge =
             transcript.challenge_scalar(b"variable base separation challenge");
+        let lookup_sep_challenge = transcript.challenge_scalar(b"lookup challenge");
 
         // Add commitment to quotient polynomial to transcript
         transcript.append_commitment(b"t_1", &self.t_1_comm);
@@ -200,14 +234,41 @@ impl Proof {
         transcript.append_commitment(b"t_3", &self.t_3_comm);
         transcript.append_commitment(b"t_4", &self.t_4_comm);
 
-        // Compute evaluation challenge
-        let z_challenge = transcript.challenge_scalar(b"z");
-
         // Compute zero polynomial evaluated at `z_challenge`
         let z_h_eval = domain.evaluate_vanishing_polynomial(&z_challenge);
 
         // Compute first lagrange polynomial evaluated at `z_challenge`
         let l1_eval = compute_first_lagrange_evaluation(&domain, &z_h_eval, &z_challenge);
+
+        // Compute n'th Lagrange poly evaluated at `z_challenge`
+        let ln_eval = compute_nth_lagrange_evaluation(&domain, &z_h_eval, &z_challenge);
+
+        // Compress table into vector of single elements
+        let mut compressed_t: Vec<BlsScalar> = lookup_table
+            .0
+            .iter()
+            .map(|arr| arr[0] + arr[1] * zeta + arr[2] * zeta * zeta + arr[3] * zeta * zeta * zeta)
+            .collect();
+
+        // Sort table so we can be sure to choose an element that is not the highest or lowest
+        compressed_t.sort();
+        let second_element = compressed_t[1];
+
+        // Pad the table to the correct size with an element that is not the highest or lowest
+        let pad = vec![second_element; domain.size() - compressed_t.len()];
+        compressed_t.extend(pad);
+
+        // Sort again to return t to sorted state
+        // There may be a better way of inserting the padding so the sort does not need to happen twice
+        compressed_t.sort();
+
+        let compressed_t_multiset = MultiSet(compressed_t);
+
+        // Compute table poly
+        let t = Polynomial::from_coefficients_vec(domain.ifft(&compressed_t_multiset.0.as_slice()));
+
+        let table_eval = t.evaluate(&z_challenge);
+        let table_next_eval = t.evaluate(&(z_challenge * domain.group_gen));
 
         // Compute quotient polynomial evaluated at `z_challenge`
         let t_eval = self.compute_quotient_evaluation(
@@ -216,10 +277,15 @@ impl Proof {
             &alpha,
             &beta,
             &gamma,
+            &delta,
+            &epsilon,
+            &zeta,
             &z_challenge,
             &z_h_eval,
             &l1_eval,
+            &ln_eval,
             &self.evaluations.perm_eval,
+            &lookup_sep_challenge,
         );
 
         // Compute commitment to quotient polynomial
@@ -241,7 +307,12 @@ impl Proof {
         transcript.append_scalar(b"q_c_eval", &self.evaluations.q_c_eval);
         transcript.append_scalar(b"q_l_eval", &self.evaluations.q_l_eval);
         transcript.append_scalar(b"q_r_eval", &self.evaluations.q_r_eval);
+        transcript.append_scalar(b"q_lookup_eval", &self.evaluations.q_lookup_eval);
         transcript.append_scalar(b"perm_eval", &self.evaluations.perm_eval);
+        transcript.append_scalar(b"lookup_perm_eval", &self.evaluations.lookup_perm_eval);
+        transcript.append_scalar(b"h_1_eval", &self.evaluations.h_1_eval);
+        transcript.append_scalar(b"h_1_next_eval", &self.evaluations.h_1_next_eval);
+        transcript.append_scalar(b"h_2_next_eval", &self.evaluations.h_2_next_eval);
         transcript.append_scalar(b"t_eval", &t_eval);
         transcript.append_scalar(b"r_eval", &self.evaluations.lin_poly_eval);
 
@@ -250,14 +321,21 @@ impl Proof {
             &alpha,
             &beta,
             &gamma,
+            &delta,
+            &epsilon,
             (
                 &range_sep_challenge,
                 &logic_sep_challenge,
                 &fixed_base_sep_challenge,
                 &var_base_sep_challenge,
+                &lookup_sep_challenge,
             ),
             &z_challenge,
             l1_eval,
+            ln_eval,
+            table_eval,
+            table_next_eval,
+            domain.group_gen_inv,
             &verifier_key,
         );
 
@@ -287,6 +365,8 @@ impl Proof {
             self.evaluations.out_sigma_eval,
             verifier_key.permutation.out_sigma,
         ));
+        aggregate_proof.add_part((self.evaluations.f_short_eval, self.f_comm));
+        aggregate_proof.add_part((self.evaluations.h_1_eval, self.h_1_comm));
         // Flatten proof with opening challenge
         let flattened_proof_a = aggregate_proof.flatten(transcript);
 
@@ -296,12 +376,13 @@ impl Proof {
         shifted_aggregate_proof.add_part((self.evaluations.a_next_eval, self.a_comm));
         shifted_aggregate_proof.add_part((self.evaluations.b_next_eval, self.b_comm));
         shifted_aggregate_proof.add_part((self.evaluations.d_next_eval, self.d_comm));
+        shifted_aggregate_proof.add_part((self.evaluations.h_1_next_eval, self.h_1_comm));
+        shifted_aggregate_proof.add_part((self.evaluations.h_2_next_eval, self.h_2_comm));
+        shifted_aggregate_proof.add_part((self.evaluations.lookup_perm_eval, self.p_comm));
         let flattened_proof_b = shifted_aggregate_proof.flatten(transcript);
-
         // Add commitment to openings to transcript
         transcript.append_commitment(b"w_z", &self.w_z_comm);
         transcript.append_commitment(b"w_z_w", &self.w_zw_comm);
-
         // Batch check
         if opening_key
             .batch_check(
@@ -313,10 +394,11 @@ impl Proof {
         {
             return Err(ProofErrors::ProofVerificationError.into());
         }
+
         Ok(())
     }
 
-    #[allow(clippy::too_many_arguments)]
+ #[allow(clippy::too_many_arguments)]
     fn compute_quotient_evaluation(
         &self,
         domain: &EvaluationDomain,
@@ -324,15 +406,37 @@ impl Proof {
         alpha: &BlsScalar,
         beta: &BlsScalar,
         gamma: &BlsScalar,
+        delta: &BlsScalar,
+        epsilon: &BlsScalar,
+        zeta: &BlsScalar,
         z_challenge: &BlsScalar,
         z_h_eval: &BlsScalar,
         l1_eval: &BlsScalar,
+        ln_eval: &BlsScalar,
         z_hat_eval: &BlsScalar,
+        lookup_sep_challenge: &BlsScalar,
     ) -> BlsScalar {
+        let omega_inv = domain.group_gen_inv;
+
         // Compute the public input polynomial evaluated at `z_challenge`
         let pi_eval = compute_barycentric_eval(pub_inputs, z_challenge, domain);
 
+        // Compute powers of alpha_0
         let alpha_sq = alpha.square();
+
+        // Compute powers of alpha_1
+        let l_sep_2 = lookup_sep_challenge.square();
+        let l_sep_3 = lookup_sep_challenge * l_sep_2;
+        let l_sep_4 = lookup_sep_challenge * l_sep_3;
+        let l_sep_5 = lookup_sep_challenge * l_sep_4;
+
+        // Compute power of zeta
+        let zeta_sq = zeta.square();
+        let zeta_cu = zeta_sq * zeta;
+
+        // Compute common term
+        let epsilon_one_plus_delta = epsilon * (BlsScalar::one() + delta);
+
         // r + PI(z)
         let a = self.evaluations.lin_poly_eval + pi_eval;
 
@@ -348,16 +452,40 @@ impl Proof {
         let beta_sig3 = beta * self.evaluations.out_sigma_eval;
         let b_2 = self.evaluations.c_eval + beta_sig3 + gamma;
 
-        // ((d + gamma) * z_hat) * alpha
+        // ((d + gamma) * z_hat) * alpha_0
         let b_3 = (self.evaluations.d_eval + gamma) * z_hat_eval * alpha;
 
         let b = b_0 * b_1 * b_2 * b_3;
 
-        // l_1(z) * alpha^2
+        // l_1(z) * alpha_0^2
         let c = l1_eval * alpha_sq;
 
+        // q_lookup(z) * (a + b*zeta + c*zeta^2 + d*zeta^3) * alpha_1
+        let d_0 = self.evaluations.a_eval
+            + (self.evaluations.b_eval * zeta)
+            + (self.evaluations.c_eval * zeta_sq)
+            + (self.evaluations.d_eval * zeta_cu);
+        let d = self.evaluations.q_lookup_eval * d_0 * lookup_sep_challenge;
+
+        // l_1(z) * alpha_1^2
+        let e = l1_eval * l_sep_2;
+
+        // (z - omega_inv) * p_eval * (epsilon( 1+ delta) + h_1_eval +(delta * h_1_next_eval)(epsilon( 1+ delta) + delta * h_2_next_eval) * alpha_1^3
+        let f_0 = z_challenge - omega_inv;
+        let f_1 = epsilon_one_plus_delta
+            + self.evaluations.h_1_eval
+            + (delta * self.evaluations.h_1_next_eval);
+        let f_2 = epsilon_one_plus_delta + (delta * self.evaluations.h_2_next_eval);
+        let f = f_0 * self.evaluations.lookup_perm_eval * f_1 * f_2 * l_sep_3;
+
+        // l_n(z) * h_2_next_eval * alpha_1^4
+        let g = ln_eval * self.evaluations.h_2_next_eval * l_sep_4;
+
+        // l_n(z) * alpha_1^5
+        let h = ln_eval * l_sep_5;
+
         // Return t_eval
-        (a - b - c) * z_h_eval.invert().unwrap()
+        (a - b - c + d - e - f - g - h) * z_h_eval.invert().unwrap()
     }
 
     fn compute_quotient_commitment(&self, z_challenge: &BlsScalar, n: usize) -> Commitment {
@@ -378,14 +506,21 @@ impl Proof {
         alpha: &BlsScalar,
         beta: &BlsScalar,
         gamma: &BlsScalar,
+        delta: &BlsScalar,
+        epsilon: &BlsScalar,
         (
             range_sep_challenge,
             logic_sep_challenge,
             fixed_base_sep_challenge,
             var_base_sep_challenge,
-        ): (&BlsScalar, &BlsScalar, &BlsScalar, &BlsScalar),
+            lookup_sep_challenge,
+        ): (&BlsScalar, &BlsScalar, &BlsScalar, &BlsScalar, &BlsScalar),
         z_challenge: &BlsScalar,
         l1_eval: BlsScalar,
+        ln_eval: BlsScalar,
+        t_eval: BlsScalar,
+        t_next_eval: BlsScalar,
+        omega_inv: BlsScalar,
         verifier_key: &VerifierKey,
     ) -> Commitment {
         let mut scalars: Vec<_> = Vec::with_capacity(6);
@@ -425,6 +560,23 @@ impl Proof {
             &self.evaluations,
         );
 
+        verifier_key.lookup.compute_linearisation_commitment(
+            lookup_sep_challenge,
+            &mut scalars,
+            &mut points,
+            &self.evaluations,
+            z_challenge,
+            (&delta, &epsilon),
+            &l1_eval,
+            &ln_eval,
+            &t_eval,
+            &t_next_eval,
+            self.h_1_comm.0,
+            self.h_2_comm.0,
+            self.p_comm.0,
+            &omega_inv,
+        );
+
         verifier_key.permutation.compute_linearisation_commitment(
             &mut scalars,
             &mut points,
@@ -446,6 +598,17 @@ fn compute_first_lagrange_evaluation(
 ) -> BlsScalar {
     let n_fr = BlsScalar::from(domain.size() as u64);
     let denom = n_fr * (z_challenge - BlsScalar::one());
+    z_h_eval * denom.invert().unwrap()
+}
+
+fn compute_nth_lagrange_evaluation(
+    domain: &EvaluationDomain,
+    z_h_eval: &BlsScalar,
+    z_challenge: &BlsScalar,
+) -> BlsScalar {
+    let n_fr = BlsScalar::from(domain.size() as u64);
+    let eval_point = z_challenge * domain.group_gen;
+    let denom = n_fr * (eval_point - BlsScalar::one());
     z_h_eval * denom.invert().unwrap()
 }
