+++ src/proof_system/widget/ecc/curve_addition/mod.rs
@@ -8,4 +8,4 @@ mod proverkey;
 mod verifierkey;
 
 pub use proverkey::ProverKey;
-pub use verifierkey::{PlookupVerifierKey, VerifierKey};
+pub use verifierkey::VerifierKey;
