+++ src/proof_system/widget/range/verifierkey.rs
@@ -7,7 +7,6 @@
 use super::delta;
 use crate::commitment_scheme::kzg10::Commitment;
 use crate::proof_system::linearisation_poly::ProofEvaluations;
-use crate::proof_system::lookup_lineariser::PlookupProofEvaluations;
 use dusk_bls12_381::{BlsScalar, G1Affine};
 
 #[derive(Debug, PartialEq, Eq, Copy, Clone)]
@@ -22,35 +21,6 @@ impl VerifierKey {
         scalars: &mut Vec<BlsScalar>,
         points: &mut Vec<G1Affine>,
         evaluations: &ProofEvaluations,
-    ) {
-        let four = BlsScalar::from(4);
-
-        let kappa = range_separation_challenge.square();
-        let kappa_sq = kappa.square();
-        let kappa_cu = kappa_sq * kappa;
-
-        let b_1 = delta(evaluations.c_eval - (four * evaluations.d_eval));
-        let b_2 = delta(evaluations.b_eval - four * evaluations.c_eval) * kappa;
-        let b_3 = delta(evaluations.a_eval - four * evaluations.b_eval) * kappa_sq;
-        let b_4 = delta(evaluations.d_next_eval - (four * evaluations.a_eval)) * kappa_cu;
-
-        scalars.push((b_1 + b_2 + b_3 + b_4) * range_separation_challenge);
-        points.push(self.q_range.0);
-    }
-}
-
-#[derive(Debug, PartialEq, Eq, Copy, Clone)]
-pub struct PlookupVerifierKey {
-    pub q_range: Commitment,
-}
-
-impl PlookupVerifierKey {
-    pub(crate) fn compute_linearisation_commitment(
-        &self,
-        range_separation_challenge: &BlsScalar,
-        scalars: &mut Vec<BlsScalar>,
-        points: &mut Vec<G1Affine>,
-        evaluations: &PlookupProofEvaluations,
     ) {
         let four = BlsScalar::from(4);
